+++ src/CPT/Main.java
@@ -1,4 +1,4 @@
-package basic;
+package CPT;
 
 public class Main{
     public static void main(String[] args){
